Cleanup - Moving interface unit tests to their own package
