+++ src/test/java/com/orasi/core/interfaces/TestTextbox.java
@@ -1,4 +1,4 @@
-package com.orasi.core;
+package com.orasi.core.interfaces;
 
 import org.openqa.selenium.By;
 import org.testng.Assert;
